backend/requirements.txt
fastapi
uvicorn
pydantic-settings
structlog
celery
redis
neo4j
faiss-cpu
sentence-transformers
tree-sitter
tree-sitter-python
tree-sitter-javascript
numpy
pytest

